+++ spec/features/staff/fund_managers_can_view_fund_level_activities_spec.rb
@@ -8,6 +8,7 @@ RSpec.feature "Fund managers can view fund level activities" do
   end
 
   let(:organisation) { create(:organisation) }
+  let!(:fund_activity) { create(:activity, level: :fund, organisation: organisation) }
 
   context "when the user is a fund_manager" do
     before do
@@ -15,12 +16,27 @@ RSpec.feature "Fund managers can view fund level activities" do
     end
 
     scenario "the user will see fund level activities on the organisation show page" do
-      fund_activity = create(:activity, level: :fund, organisation: organisation)
       visit organisations_path
       click_link organisation.name
 
       expect(page).to have_content(I18n.t("page_content.organisation.funds"))
       expect(page).to have_content fund_activity.title
+    end
+
+    scenario "can view a fund level activity" do
+      visit organisation_activity_path(fund_activity.organisation, fund_activity)
+
+      expect(page).to have_content fund_activity.title
+    end
+
+    scenario "can view and create programme level activities" do
+      programme_activity = create(:activity, level: :programme)
+      fund_activity.activities << programme_activity
+      activity_presenter = ActivityPresenter.new(programme_activity)
+      visit organisation_activity_path(fund_activity.organisation, fund_activity)
+
+      expect(page).to have_link activity_presenter.display_title
+      expect(page).to have_button I18n.t("page_content.organisation.button.create_programme")
     end
 
     context "when the activity is partially complete and doesn't have a title" do